Cleanup tests usage of filesystem
Some tests used the filesystem (bin) but didn't really need it, so I
fixed them so that they could run as unit tests now that we aren't
building before running the tests.

Other tests really do need to hit the file system and need a properly
populated bin, so I promoted them to integration tests.
