+++ pkg/mixin/provider/filesystem_test.go
@@ -1,3 +1,5 @@
+// +build integration
+
 package mixinprovider
 
 import (
