+++ pkg/porter/mixins_test.go
@@ -1,7 +1,6 @@
 package porter
 
 import (
-	"path/filepath"
 	"testing"
 
 	"github.com/deislabs/porter/pkg/mixin"
@@ -14,23 +13,12 @@ func TestPorter_PrintMixins(t *testing.T) {
 	p := NewTestPorter(t)
 	p.TestConfig.SetupPorterHome()
 
-	// Start with a fresh mixins dir to make the test more durable as we add more mixins later
-	mixinsDir, err := p.GetMixinsDir()
-	require.Nil(t, err)
-	err = p.FileSystem.RemoveAll(mixinsDir)
-	require.Nil(t, err)
-
-	// Just copy in the exec and helm mixins
-	srcMixinsDir := filepath.Join(p.TestConfig.TestContext.FindBinDir(), "mixins")
-	p.TestConfig.TestContext.AddTestDirectory(filepath.Join(srcMixinsDir, "/helm"), filepath.Join(mixinsDir, "helm"))
-	p.TestConfig.TestContext.AddTestDirectory(filepath.Join(srcMixinsDir, "/exec"), filepath.Join(mixinsDir, "exec"))
-
 	opts := PrintMixinsOptions{
 		PrintOptions: printer.PrintOptions{
 			Format: printer.FormatTable,
 		},
 	}
-	err = p.PrintMixins(opts)
+	err := p.PrintMixins(opts)
 
 	require.Nil(t, err)
 	wantOutput := "exec\nhelm\n"
